Moving whole UI test logic into test case class
